Much improved attachment handling.
MIMEParser: "Base64" parts are now called raw parts and can be encoded
using any JavaMail-supported encoding. A subclass can now decide if a
part should be forced to be a raw part.

XMIMEMessage: All but the first of an multipart/mixed section's
children are considered to be attachments, no matter what MIME type
they have, and will be stored in a XIOS document stream.